+++ packages/visual-editor/src/components/Dropzone/EditorBlock.tsx
@@ -11,8 +11,7 @@ import type {
   ResolveDesignValueType,
 } from '@contentful/experience-builder-core/types';
 import {
-  CONTENTFUL_CONTAINER_ID,
-  CONTENTFUL_SINGLE_COLUMN_ID,
+  CONTENTFUL_COMPONENTS,
   ASSEMBLY_BLOCK_NODE_TYPE,
   OUTGOING_EVENTS,
 } from '@contentful/experience-builder-core/constants';
@@ -56,6 +55,7 @@ export const EditorBlock: React.FC<EditorBlockProps> = ({
 
   const isContainer = node.data.blockId === CONTENTFUL_COMPONENTS.container.id;
   const isSingleColumn = node.data.blockId === CONTENTFUL_COMPONENTS.singleColumn.id;
+
   const containsZone = sectionsWithZone[componentId];
 
   const isAssemblyBlock = node.type === ASSEMBLY_BLOCK_NODE_TYPE;
@@ -93,7 +93,7 @@ export const EditorBlock: React.FC<EditorBlockProps> = ({
     setHoveringZone('');
   };
 
-  if (node.data.blockId === CONTENTFUL_SINGLE_COLUMN_ID) {
+  if (node.data.blockId === CONTENTFUL_COMPONENTS.singleColumn.id) {
     return (
       <DraggableChildComponent
         elementToRender={elementToRender}