+++ packages/storybook-addon/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-alpha.20](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.18...@contentful/experience-builder-storybook-addon@0.0.2-alpha.20) (2024-02-01)
+
+### Bug Fixes
+
+- allow built-in components to be disabled [ALT-267] ([#321](https://github.com/contentful/experience-builder-storybook/issues/321)) ([453225c](https://github.com/contentful/experience-builder-storybook/commit/453225cf9c0813137441e1445754c4ed3aecd8bc))
+
+### Features
+
+- section component [ALT-205] ([#279](https://github.com/contentful/experience-builder-storybook/issues/279)) ([aea2ac8](https://github.com/contentful/experience-builder-storybook/commit/aea2ac899dbf12e9e62ac174d84b5171ce60dbdb))
+
 ## [0.0.2-alpha.18](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.17...@contentful/experience-builder-storybook-addon@0.0.2-alpha.18) (2024-01-25)
 
 **Note:** Version bump only for package @contentful/experience-builder-storybook-addon